@@ -11,33 +11,41 @@ description: A new Flutter project.
 # In iOS, build-name is used as CFBundleShortVersionString while build-number used as CFBundleVersion.
 # Read more about iOS versioning at
 # https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
-version: 1.0.0+1
+version: 2.1.0+2
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
-  google_fonts: ^2.0.0
+  # updated at 1.1.0 + 1
+  get:
+  # json 
+  json_annotation:
   # auth
-  firebase_auth: ^1.1.4
+  firebase_auth: ^3.3.7
   firebase_core: "^1.0.3"
-  cloud_firestore: ^1.0.4
-  firebase_storage: ^8.0.3
+  cloud_firestore: ^3.1.8
+  firebase_storage: ^10.2.7
   # makrdown
   flutter_markdown: ^0.6.1
   # file upload
-  file_picker: ^3.0.1
-  image_picker_web: ^2.0.1
-  image_whisperer: ^0.11.1
+  file_picker: ^4.4.0
 
   flutter:
     sdk: flutter
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^0.1.2
+  cupertino_icons: ^1.0.4
+  flutter_svg: ^1.0.3
+  cached_network_image: ^3.2.0
+  image_picker: ^0.8.4+9
+  url_launcher: ^6.0.20
 
 dev_dependencies:
+  # json
+  build_runner:
+  json_serializable:
   flutter_test:
     sdk: flutter
 
@@ -53,7 +61,7 @@ flutter:
 
   # To add assets to your application, add an assets section, like this:
   assets:
-    - assets/logo.png
+    - assets/yonsei_logo.svg
   #  - images/a_dot_ham.jpeg
 
   # An image asset can refer to one or more resolution-specific "variants", see
