@@ -1,12 +1,8 @@
 package com.example.the_basics
 
-import android.os.Bundle
-import io.flutter.app.FlutterActivity
-import io.flutter.plugins.GeneratedPluginRegistrant
+// removed at 02.19 cause V2 embedding issue occurred
+import io.flutter.embedding.android.FlutterActivity
 
 class MainActivity: FlutterActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        GeneratedPluginRegistrant.registerWith(this)
-    }
+    // removed at 02.19 cause V2 embedding issue occurred
 }
